+++ src/slogo/Model/Commands/TurtleCommands/BackCommand.java
@@ -35,8 +35,8 @@ public class BackCommand extends Command {
   }
 
   @Override
-  public Integer returnArgValue() {
-    return (int) this.originalArgValue;
+  public Double returnArgValue() {
+    return this.originalArgValue;
   }
 
 }
