+++ src/slogo/Model/Commands/TurtleCommands/TowardsCommand.java
@@ -59,9 +59,9 @@ public class TowardsCommand extends Command {
   }
 
   @Override
-  public Integer returnArgValue() {
+  public Double returnArgValue() {
 
-    return (int) this.returnArgValue;
+    return this.returnArgValue;
   }
 
   private void constrainAngle(){
